Incorporates EGAP registration creation with the import_EGAP management command
